requirements.txt
PyYAML>=6.0.1
boto3>=1.20.24     # For AWS S3 integration (optional)
